Modified docdbExportUsers.py to export user & custom role information. Modified README.md to reflect the new bahavior.
